+++ js/data.js
@@ -1,3 +1,5 @@
+import {getRandomArrayElement} from './util.js';
+
 const ID = [];
 for (let i = 1; i <= 25; i++) {
   ID.push(i);
@@ -50,20 +52,6 @@ const NAME = [
 
 const QUANTITY_PHOTOS = 25;
 
-
-const getRandomInteger = (a, b) => {
-  const lower = Math.ceil(Math.min(a, b));
-  const upper = Math.floor(Math.max(a, b));
-  const result = Math.random() *
-    (
-      upper - lower + 1
-    ) +
-    lower;
-  return Math.floor(result);
-};
-
-const getRandomArrayElement = (elements) => elements[getRandomInteger(0, elements.length - 1)];
-
 const createObject = () => (
   {
     id: getRandomArrayElement(ID),
@@ -79,4 +67,5 @@ const createObject = () => (
 
 const similarObject = Array.from({ length: QUANTITY_PHOTOS }, createObject);
 
-// console.log(similarObject);
+console.log(similarObject);
+
