+++ package.json
@@ -20,5 +20,7 @@
   "scripts": {
     "start": "browser-sync start --server --no-ui --files \"js/**/*.js\"",
     "lint": "eslint js/"
-  }
+  },
+  "type": "module"
 }
+
